fix: Passport 인증 오류 수정 (UUID 및 타입 캐스팅)
- OAuth 테이블 user_id 컬럼을 UUID 타입으로 변경
- AuthService에서 Carbon 메서드 int 타입 캐스팅 추가
- User 모델에 OAuthenticatable 인터페이스 구현

+++ app/Services/AuthService.php
@@ -131,7 +131,7 @@ class AuthService
             'id' => bin2hex(random_bytes(40)),
             'access_token_id' => $accessToken->id,
             'revoked' => false,
-            'expires_at' => now()->addDays(config('passport.refresh_tokens_expire_in', 14)),
+            'expires_at' => now()->addDays((int) config('passport.refresh_tokens_expire_in', 14)),
         ]);
 
         return $refreshToken->id;
